requirements.txt
numerize
pandas
plotly-express
streamlit